content/base/test/test_bug444030.xhtml

<?xml version="1.0"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=444030
-->
<head>
  <title>Test for Bug 444030</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <bindings xmlns="http://www.mozilla.org/xbl">
    <binding id="e">
      <content>e</content>
    </binding>
  </bindings>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=444030">Mozilla Bug 444030</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

/** Test for Bug 444030 **/

function doTest() {
  var anonTextNode = document.getAnonymousNodes(document.getElementById("boundElement"))[0];
  var hadException = false;
  try {
    var wholeText = anonTextNode.wholeText;
  } catch(e) {
    hadException = true;
  }
  ok(hadException,
     "Should have got an exception when using .wholeText with a text node child of the binding parent");

  hadException = false;
  try {
    anonTextNode.replaceWholeText("foobar");
  } catch(e) {
    hadException = true;
  }
  ok(hadException,
     "Should have got an exception when using .replaceWholeText with a text node child of the binding parent");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(doTest);

]]>
</script>
</pre>
<div style="-moz-binding: url(#e)" id="boundElement"></div>
</body>
</html>
